https://github.com/nehasri5521/CAPSTONE--1---EDA-GOOGLE-PLAY-STORE-DATA/upload
The objective of this project is to deliver insights to understand customer demands better and thus help developers to popularize the product. It is of 10k Play Store apps data for analyzing the Android market. This dataset contains details of different applications and reviews from different users.
The Play Store apps data has enormous potential to drive app-making businesses to success. Actionable insights can be drawn for developers to work on and capture the Android market. Each app (row) has values for category, rating, size, and more. Another dataset contains customer reviews of the android apps. Explore and analyse the data to discover key factors responsible for app engagement and success.